Added entity manager to remove

+++ src/Controller/Api/FileController.php
@@ -6,6 +6,7 @@ use App\Entity\File;
 use App\Message\DeleteFile;
 use App\Repository\FileRepository;
 
+use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 
 use Symfony\Component\HttpFoundation\Response;
@@ -21,23 +22,25 @@ class FileController extends AbstractFOSRestController
     /**
      * Delete a Dataset and associated Metadata and Difs.
      *
-     * @param string              $id             The id of the File to delete.
-     * @param FileRepository      $fileRepository File entity Repository instance.
-     * @param MessageBusInterface $messageBus     Symfony messenger bus interface instance.
-     *
+     * @param string                 $id             The id of the File to delete.
+     * @param FileRepository         $fileRepository File entity Repository instance.
+     * @param MessageBusInterface    $messageBus     Symfony messenger bus interface instance.
+     * @param EntityManagerInterface $entityManager  Entity manager interface instance.
      * @Route("/api/file/{id}", name="pelagos_api_datasets_delete", methods={"DELETE"}, defaults={"_format"="json"})
      *
      * @return Response A response object with an empty body and a "no content" status code.
      */
-    public function deleteFile(string $id, FileRepository $fileRepository, MessageBusInterface $messageBus)
+    public function deleteFile(string $id, FileRepository $fileRepository, MessageBusInterface $messageBus, EntityManagerInterface $entityManager)
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $file = $fileRepository->find((int)$id);
 
         if ($file instanceof File) {
             $filePath = $file->getFilePath();
             $deleteFileMessage = new DeleteFile($filePath);
             $messageBus->dispatch($deleteFileMessage);
-            $this->handleDelete(File::class, (int)$id);
+            $entityManager->remove($file);
+            $entityManager->flush();
         } else {
             throw new BadRequestHttpException('File does not exist');
         }
